requirements.txt
aiohttp~=3.11.10
anthropic~=0.29.0
azure-core~=1.32.0
azure-cosmos==4.7.0
azure-identity~=1.19.0
azure-keyvault==4.2.0
azure-storage-blob==12.20.0
beautifulsoup4~=4.12.3
cohere~=5.11.3
fastapi~=0.115.8
google-generativeai==0.6.0
html2text~=2024.2.26
instructor~=1.7.2
msal~=1.31.1
nltk~=3.9.1
openai~=1.59.7
pathlib~=1.0.1
playwright~=1.49.0
psycopg2==2.9.10
pydantic~=2.8.2
pydantic-settings==2.4.0
PyPDF2~=3.0.1
pymupdf4llm~=0.0.17
python-dotenv==1.0.0
requests~=2.32.3
slack_sdk~=3.33.4
stripe==7.13.0
tiktoken >= 0.7
tqdm~=4.67.1
uvicorn~=0.34.0
ddgs~=9.5.2
email-validator~=2.2.0